Rename ChunkedArray -> chunked_array files
